* new savegame format - this is currently work-in-progress and breaks all old savegames NOTE: Every subsystem has its own save and load function (see cl_save.h for their names)

+++ src/client/cl_research.h
@@ -151,7 +151,7 @@ typedef struct technology_s {
 	char mdl_top[MAX_QPATH];
 	char mdl_bottom[MAX_QPATH];
 
-	byte statusResearchable;		/**< Is this item researchable? */
+	int statusResearchable;		/**< Is this item researchable? */
 
 	int produceTime;			/**< How many days the production of this items runs. */
 	qboolean pushnews;			/**< Push news about this tech when researched. Note: usually you should use this only for tech with time = 0. */
